+++ packages/assignment-7/src/utils/getWeeksInMonth.ts
@@ -1,0 +1,24 @@
+import { getDaysInMonth } from "./getDaysInMonth";
+
+export const getWeeksInMonth = (year: number, month: number) => {
+  const daysInMonth = getDaysInMonth(year, month);
+  const firstDayOfMonth = new Date(year, month, 1).getDay();
+  const days = Array.from({ length: daysInMonth }, (_, i) => i + 1);
+  const weeks = [];
+  let week = Array(7).fill(null);
+
+  for (let i = 0; i < firstDayOfMonth; i++) {
+    week[i] = null;
+  }
+
+  for (const day of days) {
+    const dayIndex = (firstDayOfMonth + day - 1) % 7;
+    week[dayIndex] = day;
+    if (dayIndex === 6 || day === daysInMonth) {
+      weeks.push(week);
+      week = Array(7).fill(null);
+    }
+  }
+
+  return weeks;
+};
